renaming model,control and validationSchema
