+++ controller/userController.js
@@ -1,10 +1,12 @@
-const UserModel = require('../model/user');
+const UserModel = require('../model/userModel');
 // const { SendMail } = require('../utils/SendMail');
-exports.signUpUser = async (req, res) => {
+exports.signUpUsr = async (req, res) => {
   try {
     registration_data = req.body;
 
     await UserModel.createUser(registration_data);
+
+    //send email here response:250 = success or response:OK= success
 
     return res.status(200).json({
       status: 'success',
@@ -14,7 +16,7 @@ exports.signUpUser = async (req, res) => {
     // console.log(error);
     return res.status(404).json({
       status: 'fail',
-      message: 'error',
+      message: 'You Entered incorrect data or something unexpected happened',
       errorMessage: error.message,
     });
   }
