+++ route/userRoutes.js
@@ -1,9 +1,9 @@
 const express = require('express');
-const controller = require('../controller/users');
+const controller = require('../controller/userController');
 const validation = require('../validations/userValidations');
 
 const router = express.Router();
 
-router.post('/register', validation.signupValidation, controller.signUpUser);
+router.post('/register', validation.signupValidation, controller.signUpUsr);
 
 module.exports = router;
